Move Iso to Riemannian optimizers
