+++ docs/apidocs/riemannian-optimizers.md
@@ -8,6 +8,6 @@ emerging_optimizers.riemannian_optimizers
 .. automodule:: emerging_optimizers.riemannian_optimizers.normalized_optimizer
     :members:
 
-.. automodule:: emerging_optimizers.rl_optimizers.isospectral
+.. automodule:: emerging_optimizers.riemannian_optimizers.isospectral
     :members:
 ```
