+++ tests/test_isospectral.py
@@ -19,7 +19,7 @@ from absl import flags, logging
 from absl.testing import absltest, parameterized
 
 from emerging_optimizers import registry
-from emerging_optimizers.rl_optimizers.isospectral import Iso
+from emerging_optimizers.riemannian_optimizers.isospectral import Iso
 
 
 flags.DEFINE_enum("device", "cpu", ["cpu", "cuda"], "Device to run tests on")
